[Perseverance] Basic Playability
Added a prefab for turtles and two scripts. There are 30 turtles, one of which is the player, and only one of them makes it to the ocean. Infinite retries, but right now I think it feels more like failing constantly until you don't rather than succeeding if you don't give up.
NOTE: This commit includes some changes to shaders and debug that I didn't do myself. May be a version/operating system difference. If there are issues for others after this then that might be the cause.