Rhythm Game: Life Indicators, Song Ends
Added three circle sprites above the fret to indicate how many lives the player has in phase 2. Also added a value to the conductor for the song's length in beats, and a value to the rhythm script for a song's length in notes. The score value has been re-purposed to keep track of the song's progress. When score reaches the song length, the song ends and the rhythm game closes.