Added Fidel's Song, Improved Song End
The Proof of Concept Demo now uses Fidel's song, courtesy of Luca. It was able to be implemented with only a few public values being changed, so it should be easy to add new songs past that. Also cleaned up the end of the song, so the rhythm game elements shrink out instead of just disappearing.