Added Scripts for Overworld Camera and NPC Motion
The camera now follows the player, at least in the Proof of Concept scene. Right now the yarnspinner example camera looks like it should work for what we want. IMPORTANT: At the moment the code demands that the rhythm game should take place at the 0,0,0 position of the scene's space. We could just design all the scenes around this, but if that's inconvenient I'll revisit this.
Also, added a script that allows NPCs to move between waypoints through yarnspinner dialogue. NPCs move and can't be spoken to until they reach a destination.
For a brief, glorious moment, Fidel could teleport.