Environmental Puzzle Framework
Added scripts and prefabs for environmental puzzles using the same note system. Currently the player presses a key to activate a version of the rhythm game that just responds to note presses, and causes nearby note blocks to react to the notes you press. Currently the two kinds of note blocks are simple blocks, which switch between two positions when their indicated note is pressed, and locks, which have a combination of three note blocks that must be activated in sequence to open a path.