Implemented Puzzle, Fixed Doors, Improved Stage Directions
Moved the environment puzzle to the main proof of concept scene and implemented the puzzle art assets. Adjusted waypoints to compensate for the shifted backgrounds. Retagged the player as the player, which fixed doors, while also setting up doors so that they can only be used after a yarn command enables them. Fixed how dialogue worked with the environment puzzle and split the puzzle scene into one for before the puzzle and one for after it's solved. Various small fixes to waypoint coordinates, sprites having unlit materials, and stopping the player from colliding with puzzle objects.